requirements.txt
amqp==5.1.1
arabic-reshaper==2.1.3
asgiref==3.5.0
asn1crypto==1.5.1
async-timeout==4.0.2
bcrypt==3.2.0
billiard==3.6.4.0
black==22.3.0
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.2
cssselect2==0.6.0
Deprecated==1.2.13
distlib==0.3.4
dj-static==0.0.6
Django==4.0.4
django-countries==7.3.2
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-model-utils==4.2.0
django-ninja==0.17.0
django-phonenumber-field==6.1.0
filelock==3.6.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
identify==2.4.12
idna==3.3
importlib-metadata==4.11.3
install==1.3.5
kombu==5.2.4
lxml==4.8.0
Markdown==3.3.6
mypy-extensions==0.4.3
nodeenv==1.6.0
oscrypto==1.3.0
packaging==21.3
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.1
prompt-toolkit==3.0.29
psycopg2==2.9.3
psycopg2-binary==2.9.3
Pycco==0.6.0
pycparser==2.21
pydantic==1.9.0
Pygments==2.12.0
pyHanko==0.13.0
pyhanko-certvalidator==0.19.5
PyJWT==2.3.0
pyparsing==3.0.8
PyPDF2==1.26.0
PyPDF3==1.0.6
pystache==0.6.0
python-barcode==0.13.1
python-bidi==0.4.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
redis==4.2.2
reportlab==3.6.6
requests==2.27.1
semantic-version==2.9.0
setuptools-rust==1.2.0
six==1.16.0
smartypants==2.0.1
sqlparse==0.4.2
static3==0.7.0
svglib==1.2.1
tinycss2==1.1.1
tokenizers==0.11.6
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.9
vine==5.0.0
virtualenv==20.14.0
wcwidth==0.2.5
webencodings==0.5.1
Whoosh==2.7.4
wrapt==1.14.0
xhtml2pdf==0.2.7
zipp==3.8.0
